requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyotp==2.9.0
python-dotenv==1.0.1